requirements.txt
streamlit
langchain==0.0.354
langchain-community
langchain-text-splitters
langchain-google-genai
google-generativeai
faiss-cpu
PyPDF2
python-dotenv
